separate kebab menu into a single component
